browser-sdk: Replace storybook webpack with vite
