--- packages/browser-sdk/.babelrc.json
@@ -1,19 +1,0 @@
-{
-    "sourceType": "unambiguous",
-    "presets": [
-        [
-            "@babel/preset-env",
-            {
-                "targets": {
-                    "chrome": 100,
-                    "safari": 15,
-                    "firefox": 91
-                }
-            }
-        ],
-        "@babel/preset-typescript",
-        "@babel/preset-react"
-    ],
-    "plugins": []
-}
-
